Refactorings and documentation.
